requirements.txt
distribute==0.6.24
wsgiref==0.1.2
RPi.GPIO==0.2.0
